Rename badly named file defining crag filters.
